Configure the api, add rest tests, fix item.cs
