+++ src/api/ruby.drum.Api/tests/rest/catalog.http
@@ -1,0 +1,10 @@
+### Should return 200
+GET http://localhost:5263/api/catalog HTTP/1.1
+
+
+### Should return 200
+GET http://localhost:5263/api/catalog/42 HTTP/1.1
+
+
+### Should return 200
+GET http://localhost:5263/api/catalog/test HTTP/1.1
