+++ src/api/ruby.drum.Domain/Catalog/Item.cs
@@ -42,27 +42,3 @@ public class Item
         Ratings.Add(rating);
     }
 }
-
-public class Rating
-{
-    public int Stars { get; set;}
-    public string UserName { get; set;}
-    public string? Review { get; set;}
-
-    public Rating(int stars, string userName, string review)
-    {
-        if(stars < 1 || stars > 5)
-        {
-            throw new ArgumentException("Stars must be between 1 and 5");
-        }
-
-        if(string.IsNullOrWhiteSpace(userName))
-        {
-            throw new ArgumentException("User name is required");
-        }
-
-        Stars = stars;
-        UserName = userName;
-        Review = review;
-    }
-}
